requirements.txt
certifi==2018.4.16
Django==2.0.2
django-mathfilters==0.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
Pillow==5.0.0
pytz==2018.3
six==1.11.0
wincertstore==0.2
